# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_keepguard_bff-auth_internal_adapters_inbound_http_dto.SendResetPasswordMessageRequestDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_keepguard_bff-auth_internal_adapters_inbound_http_dto.SendResetPasswordMessageResponseDTO:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_keepguard_bff-auth_internal_pkg.ErrorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_keepguard_bff-auth_internal_pkg.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/github_com_keepguard_bff-auth_internal_pkg.ErrorDetail'
        type: array
      error:
        type: string
      message:
        type: string
      traceId:
        type: string
    type: object
host: localhost:8381
info:
  contact: {}
  description: Backend for Frontend responsável pela autenticação e autorização do
    sistema KeepGuard
  title: BFF-AUTH API
  version: 1.0.0
paths:
...
  /message/reset-password/send:
    post:
      consumes:
      - application/json
      description: Envia uma mensagem de reset de senha para o email do usuário. Não
        requer autenticação. Requer headers obrigatórios X-Correlation-ID e X-Tenant-Id.
      parameters:
      - description: ID de correlação para rastreamento da requisição
        in: header
        name: X-Correlation-ID
        required: true
        type: string
      - description: ID da aplicação cliente (UUID)
        in: header
        name: X-Tenant-Id
        required: true
        type: string
      - description: Email do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_keepguard_bff-auth_internal_adapters_inbound_http_dto.SendResetPasswordMessageRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Mensagem enviada com sucesso
          schema:
            $ref: '#/definitions/github_com_keepguard_bff-auth_internal_adapters_inbound_http_dto.SendResetPasswordMessageResponseDTO'
        "400":
          description: Erro de validação (headers ausentes, email inválido ou usuário
            não ativo)
          schema:
            $ref: '#/definitions/github_com_keepguard_bff-auth_internal_pkg.ErrorResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/github_com_keepguard_bff-auth_internal_pkg.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/github_com_keepguard_bff-auth_internal_pkg.ErrorResponse'
      summary: Send reset password message
      tags:
      - messages
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"